Extract Module: RDocRuboCop::Lang::Ruby
