--- lib/rdoc_rubocop/comment_token_organizable.rb
@@ -1,17 +1,0 @@
-module RDocRuboCop
-  module CommentTokenOrganizable
-    def self.included(base)
-      base.include InstanceMethods
-
-      attr_reader :comment_tokens
-    end
-
-    module InstanceMethods
-      private
-
-      def comment_indent
-        @comment_indent ||= @comment_tokens.reject(&:blank?).map(&:comment_indent).min
-      end
-    end
-  end
-end
