+++ lib/rdoc_rubocop/lang.rb
@@ -1,0 +1,6 @@
+module RDocRuboCop
+  module Lang
+  end
+end
+
+require "rdoc_rubocop/lang/ruby"
