+++ lib/rdoc_rubocop/lang/ruby.rb
@@ -1,0 +1,8 @@
+module RDocRuboCop
+  module Lang
+    module Ruby
+    end
+  end
+end
+
+require "rdoc_rubocop/lang/ruby/source_file"
