+++ lib/rdoc_rubocop/source_file.rb
@@ -1,7 +1,5 @@
 require "digest"
-require "rdoc_rubocop/comment_extractor"
 require "rdoc_rubocop/file_path"
-require "rdoc_rubocop/source_file/corrector"
 
 module RDocRuboCop
   class SourceFile
@@ -9,8 +7,15 @@ module RDocRuboCop
     attr_reader :filename
 
     def self.build(filename)
+      klass =
+        case filename
+        when /\.rb\z/ then Lang::Ruby::SourceFile
+        end
+
+      return unless klass
+
       source = File.open(filename, "r").read
-      new(source, filename)
+      klass.new(source, filename)
     end
 
     def initialize(source, filename)
@@ -28,10 +33,14 @@ module RDocRuboCop
     end
 
     def comments
-      comment_extractor = CommentExtractor.new(self)
+      comment_extractor = comment_extractor_class.new(self)
       comment_extractor.extract
       comment_extractor.comments
     end
+
+    # def comment_extractor_class
+    #   CommentExtractor
+    # end
 
     def correct!
       correct
@@ -40,11 +49,15 @@ module RDocRuboCop
     end
 
     def correct
-      corrector = Corrector.new(@source, @source_code_file_paths)
+      corrector = corrector_class.new(@source, @source_code_file_paths)
       corrector.correct
 
       @source = corrector.source
     end
+
+    # def corrector_class
+    #   Corrector
+    # end
 
     private
 
@@ -68,3 +81,5 @@ module RDocRuboCop
     end
   end
 end
+
+require "rdoc_rubocop/lang/ruby/source_file"
