+++ spec/rdoc_rubocop/lang/ruby/comment_extractor_spec.rb
@@ -1,8 +1,8 @@
 require "spec_helper"
 
-RSpec.describe RDocRuboCop::CommentExtractor do
+RSpec.describe RDocRuboCop::Lang::Ruby::CommentExtractor do
   describe "#extract" do
-    let(:source_file) { RDocRuboCop::SourceFile.new(source, "foo.rb") }
+    let(:source_file) { RDocRuboCop::Lang::Ruby::SourceFile.new(source, "foo.rb") }
 
     subject { described_class.new(source_file).extract }
 
@@ -28,23 +28,23 @@ RSpec.describe RDocRuboCop::CommentExtractor do
       it "should returns two instances of Comment" do
         expect(subject.size).to eq(2)
 
-        expect(subject[0].comment_tokens[0]).to be_an_instance_of(RDocRuboCop::Token::CommentToken)
+        expect(subject[0].comment_tokens[0]).to be_an_instance_of(RDocRuboCop::Lang::Ruby::Token::CommentToken)
         expect(subject[0].comment_tokens[0].lineno).to eq(1)
         expect(subject[0].comment_tokens[0].token).to eq("# comment1-1\n")
-        expect(subject[0].comment_tokens[1]).to be_an_instance_of(RDocRuboCop::Token::CommentToken)
+        expect(subject[0].comment_tokens[1]).to be_an_instance_of(RDocRuboCop::Lang::Ruby::Token::CommentToken)
         expect(subject[0].comment_tokens[1].lineno).to eq(2)
         expect(subject[0].comment_tokens[1].token).to eq("# comment1-2\n")
-        expect(subject[0].comment_tokens[2]).to be_an_instance_of(RDocRuboCop::Token::CommentToken)
+        expect(subject[0].comment_tokens[2]).to be_an_instance_of(RDocRuboCop::Lang::Ruby::Token::CommentToken)
         expect(subject[0].comment_tokens[2].lineno).to eq(3)
         expect(subject[0].comment_tokens[2].token).to eq("# comment1-3\n")
 
-        expect(subject[1].comment_tokens[0]).to be_an_instance_of(RDocRuboCop::Token::CommentToken)
+        expect(subject[1].comment_tokens[0]).to be_an_instance_of(RDocRuboCop::Lang::Ruby::Token::CommentToken)
         expect(subject[1].comment_tokens[0].lineno).to eq(6)
         expect(subject[1].comment_tokens[0].token).to eq("# comment-in-Foo-1\n")
-        expect(subject[1].comment_tokens[1]).to be_an_instance_of(RDocRuboCop::Token::CommentToken)
+        expect(subject[1].comment_tokens[1]).to be_an_instance_of(RDocRuboCop::Lang::Ruby::Token::CommentToken)
         expect(subject[1].comment_tokens[1].lineno).to eq(7)
         expect(subject[1].comment_tokens[1].token).to eq("# comment-in-Foo-2\n")
-        expect(subject[1].comment_tokens[2]).to be_an_instance_of(RDocRuboCop::Token::CommentToken)
+        expect(subject[1].comment_tokens[2]).to be_an_instance_of(RDocRuboCop::Lang::Ruby::Token::CommentToken)
         expect(subject[1].comment_tokens[2].lineno).to eq(8)
         expect(subject[1].comment_tokens[2].token).to eq("# comment-in-Foo-3\n")
       end
@@ -68,15 +68,15 @@ RSpec.describe RDocRuboCop::CommentExtractor do
         expect(subject.size).to eq(2)
 
         expect(subject[0].comment_tokens.size).to eq(2)
-        expect(subject[0].comment_tokens[0]).to be_an_instance_of(RDocRuboCop::Token::CommentToken)
+        expect(subject[0].comment_tokens[0]).to be_an_instance_of(RDocRuboCop::Lang::Ruby::Token::CommentToken)
         expect(subject[0].comment_tokens[0].lineno).to eq(2)
         expect(subject[0].comment_tokens[0].token).to eq("#--\n")
-        expect(subject[0].comment_tokens[1]).to be_an_instance_of(RDocRuboCop::Token::CommentToken)
+        expect(subject[0].comment_tokens[1]).to be_an_instance_of(RDocRuboCop::Lang::Ruby::Token::CommentToken)
         expect(subject[0].comment_tokens[1].lineno).to eq(3)
         expect(subject[0].comment_tokens[1].token).to eq("# comment\n")
 
         expect(subject[1].comment_tokens.size).to eq(1)
-        expect(subject[1].comment_tokens[0]).to be_an_instance_of(RDocRuboCop::Token::CommentToken)
+        expect(subject[1].comment_tokens[0]).to be_an_instance_of(RDocRuboCop::Lang::Ruby::Token::CommentToken)
         expect(subject[1].comment_tokens[0].lineno).to eq(4)
         expect(subject[1].comment_tokens[0].token).to eq("# :nodoc:\n")
       end
