+++ spec/rdoc_rubocop/rubocop_modifier_spec.rb
@@ -3,8 +3,8 @@ require "spec_helper"
 RSpec.describe RDocRuboCop::RuboCopModifier do
   describe "#run" do
     let(:file_path) do
-      source_code = RDocRuboCop::Comment::SourceCode.new(
-        [RDocRuboCop::Token::CommentToken.new([1, 1], :on_comment, "#   [1 , 2]\n", Ripper::EXPR_BEG)],
+      source_code = RDocRuboCop::Lang::Ruby::SourceCode.new(
+        [RDocRuboCop::Lang::Ruby::Token::CommentToken.new([1, 1], :on_comment, "#   [1 , 2]\n", Ripper::EXPR_BEG)],
       )
       RDocRuboCop::FilePath.new("sample.rb", source_code)
     end
